Updated scaffolding and templating system so that the scaffolding appears within the defined template for the site.  This is dynamic so that a new page is not needed for each page, reducing maintenance and programming time.  It assumes if a request comes from say, /scaffolding/AddUsersPage that the following is true:
1) use the default scaffolding page defined in AppSettings under Global->Scaffolding->Page
2) the action is "add"
3) the tables is "Users"

if, at the end of the url is /1 then scaffolding will assume that is the ID of the object to look it up and will
automatically find the id.  So, to delete a user, use the following url "/scaffolding/DeleteUsersPage/1" 
which will show a form to delete a user with an id of 1.